--- src/polyfills/console.js
@@ -1,7 +1,0 @@
-// Console is not defined in ECMAScript, so just in case...
-module.exports = global.console || {
-  debug: function () {},
-  log: function () {},
-  warn: function () {},
-  error: function () {}
-};
